modify test module
