+++ data/toy/test_df.csv
@@ -1,0 +1,3 @@
+AGE,FRW,SBP,DBP,CHOL,CIG,sex,disease
+50,108,140,86,244,0,Male,0
+45,117,120,90,214,0,Female,0
