+++ data/toy/test_framingham.csv
@@ -1,0 +1,11 @@
+AGE,FRW,SBP,DBP,CHOL,CIG,sex,disease
+45,117,120,90,214,0,Female,0
+56,95,140,80,166,20,Male,0
+47,101,120,85,167,0,Female,0
+52,100,160,100,294,0,Female,0
+50,108,140,86,244,0,Male,0
+52,102,120,78,318,0,Female,0
+50,88,140,95,248,0,Female,0
+52,120,136,84,220,20,Female,0
+60,98,152,88,142,0,Male,0
+49,109,165,105,178,5,Female,0
