+++ data/toy/train_df.csv
@@ -1,0 +1,9 @@
+AGE,FRW,SBP,DBP,CHOL,CIG,sex_Male
+-0.06523280730534423,2.0578065752724592,-0.35901335102399884,-0.5936260127031839,0.05572437921344977,1.7002089231955175,0.0
+-0.06523280730534423,0.0419960525565808,-1.3802068828255953,-1.2562783059532499,1.6737952423002875,-0.6652991438591156,0.0
+2.022217026465671,-0.40596184138028113,0.6621801807775978,-0.1518578172031401,-1.2321279403862782,-0.6652991438591156,1.0
+-0.06523280730534423,-0.18198289441185014,1.1727769466783962,1.1734467692969917,1.2775329901157557,-0.6652991438591156,0.0
+0.9784921095801633,-0.7419302618329275,-0.10371496807359966,-1.035394208203228,-0.8358656882017466,1.7002089231955175,1.0
+-0.587095265748098,-1.5258565762224359,-0.10371496807359966,0.6212365249219367,0.5180303400954034,-0.6652991438591156,0.0
+-0.8480264949694749,0.8259223669460891,1.4918999253663952,1.7256570136720464,-0.6377345621094807,-0.07392212709545729,0.0
+-1.3698889534122287,-0.06999342092763468,-1.3802068828255953,-0.483183963828173,-0.819354761027391,-0.6652991438591156,0.0
